fix(popover): set initial top, left position
In rtl layout, popover positioned outside of viewport if 'left' set to 'auto'.
In such case, content could shrink, so we'll calculate position based on dimentions,
that would differ from those, that we'll have when it rendered inside a viewport.
With 'left' equals to '0' popover should always feet viewport and dimentions will not change.
Details:
https://www.w3.org/TR/CSS2/visuren.html#propdef-left
https://www.w3.org/TR/CSS2/visudet.html#abs-non-replaced-width