fix(menu): trigger change detection after setting max-height
To animate menu expanding we need to know height of all expoanded children.
Since children cann't be accessed before AfterViewInit hook, we have to calc
and set 'maxHeight' there, but it causing ExpressionChangedAfterItHasBeenCheckedError
error. Since 'maxHeight' can't be calculated earlier, we have to trigger change detection
again. This is a common workaround for this error.