fix(menu): trigger change detection after setting max-height (#349)
Closes #263 
To animate menu expanding we need to know a height of all expanded children.
Since children can't be accessed before AfterViewInit hook, we have to calc
and set 'maxHeight' there, but it causing ExpressionChangedAfterItHasBeenCheckedError
error. Since 'maxHeight' can't be calculated earlier, we have to trigger change detection
again. This is a common workaround for this error.