refactor(rtl): add ability to append selector to dir attr
Due to large amount of bugs with :host/:host-context selectors in angular,
we have to add [dir] attribute on body. Since styles from _component.theme.scss,
nested in .nb-theme-[theme-name], which also set on body. Therefore triсk with
'[dir=rtl] &' will lead to .nb-theme-[theme-name] [dir=rtl], so selector newer
work (since theme class and dir attribute are set on the same element).
To fix it, we have to combine [dir] with .nb-theme using 'selector-append' function.
Then result looks like this. [dir].nb-theme-[theme-name].
But on the other hand, styles from component.scss files scoped to :host, so here
we can't append, because then we have to apply dir attribute to every component with
rtl styles. So for this files we have to keep next approach:
:host {
  [dir=rtl] & {
    ...
  }
}
and resulting selector will be:
[dir=rtl] [host-attr] {
  ...
}