fix(calendar): prevent selected day text color override
Text color of the selected day from the bounding month was overridden by
a color specified in `.bounding-month`. It happened because of
`.selected` and `.bounding-month` has the same weight and the latter was
placed after former. Moving `.bounding-month` at a very beginning allows
other styles with the same weight override `.bounding-month`.