fix(tree grid): retrieve row and cell data from context
The previous approach relied on content children to get row index
and retrieve associated data by this index. It caused expression
changed exception as when new data arrives and table render cells,
cells ask a table for their level to set proper padding but content
quires not updated yet.

Instead of relying on the index element index in DOM and getting
data by this index we now use context associated with rows and cells.
Context has data and dataIndex properties so we don't need to search
for it.