fix(calendar range): handle unset date and selectedValue
In the `NbCalendarRangeDayCellComponent` `date` property could be
set to `null` in case it's from bounding month and user set
`boundingMonth` to false.
In all `NbCalendarRange*CellComponent` selectedValue could be defined
but has no `start` property set. It happens because users are setting
`[range]` input on `NbCalendarRangeComponent` to `{}` so `selectedValue`
is truthy.