build(docs): ignore example files from assets dir in prod mode
Angular 12 started to process `scss` files from the `assets` directory
as a project files when building with production configuration. To prevent
build errors we ignore `assets/examples` directory in the production
configuration and than copying it manually to the dist directory when
build is over. This way we prevent build error, but now we can't run
`ng serve docs` with production configuration as we can't copy add examples
directory to the internal Angular server. As a workaround we could build docs
and then serve it with a simple http server, for example `npx http-server docs/dist` and navigate to `localhost:8080/nebular`.
As for now it's still possible to run `ng serve docs` in a dev configuration.
Details on assets processing:
https://stackoverflow.com/a/67742437/9334725