feat(NbDialogService): support passing values to input/model signals
This makes an exception for input and model signals in the DialogConfig
context, so the typehint will show as the type of the signal's value,
but internally it will create a signal with the passed value.
Technically these will just be readonly signals instead of input or model
signals, but since components used as a dialog do not support regular
outputs anyway, it should not matter. When used in a regular context the
input signal will not be overwritten and work as usual.

closes https://github.com/akveo/nebular/issues/3256
(I would not classify this as a bug but a missing feature instead)